Cleanup 1.17 entity rewriter, fix DataItem amount check
